requirements.txt
fastapi
uvicorn
libsql-client
google-genai
python-dotenv
fastapi
uvicorn
libsql-client
google-genai
python-dotenv
httpx
mangum
